# raffleease-openapi.yaml
openapi: 3.1.0
info:
  title: Raffleease API
  description: |
    Raffleease is a RESTful API designed to support the management of online raffles for animal rescue associations. The platform follows a **multi-tenant architecture**, where most endpoints are scoped under a specific association ID to ensure data isolation across organizations.
    
    The API implements **JWT-based authentication** with secure refresh token handling via **HTTP-only, same-site, secure cookies**. **Role-based access control** governs permissions across different user types within each association.
    
    Search endpoints support **advanced filtering, sorting, and pagination** mechanisms to allow flexible querying of resources. All API responses follow a **consistent, standardized format**, and **error handling** is uniform across the platform to provide clarity for integrators.
    
    This documentation should clearly outline available endpoints, authentication flows, security requirements, and expected request and response formats to help developers integrate with the Raffleease platform effectively.    

    ## Key Features
    
    ### Multi-Tenant Architecture
    Most endpoints are scoped under `/associations/{associationId}` to ensure data isolation and support multiple organizations.
    
    ### JWT-Based Authentication
    The API uses JWT access tokens for authentication and secure HTTP-only cookies to manage refresh tokens.
    
    ### Role-Based Access Control
...
    
...
    
...
    
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new association and admin user
      description: |
        Creates a new association with an admin user account. This is the primary registration endpoint for new organizations joining the platform.
        
        **Process:**
        1. Creates new association with provided details
        2. Creates admin user account with ADMIN role
        3. Creates association membership linking user to association
        4. Sends email verification link to the user
        
        **Note:** Only admin users go through this registration process. Regular users are created by admins through the user management endpoints.
      operationId: register
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful - verification email sent
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RegisterResponse'
              example:
                success: true
                message: "New association account created successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  id: 123
                  email: "admin@animalrescue.org"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: |
        Verifies a user's email address using the verification token sent during registration.
        Users must verify their email before they can successfully log in to the platform.
      operationId: verifyEmail
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
      responses:
        '200':
          description: Email verified successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: "null"
              example:
                success: true
                message: "Account verified successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: |
        Authenticates a user with their credentials and returns an access token for API access.
        
        **Authentication Flow:**
        1. Validates user credentials (email/username + password)
        2. Verifies user account is enabled and verified
        3. Generates JWT access token
        4. Sets secure refresh token cookie
        5. Returns access token and user/association information
        
        **Identifier:** Can be either email address or username.
      operationId: login
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
            Set-Cookie:
              description: Secure refresh token cookie
              schema:
                type: string
                example: "refresh_token=eyJhbGciOiJIUzI1NiJ9...; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=2592000"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
              example:
                success: true
                message: "User authenticated successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  accessToken: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjMiLCJpYXQiOjE2MDk0NTkyMDB9..."
                  associationId: 456
                  userId: 123
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: |
        Logs out the authenticated user by invalidating their tokens and clearing the refresh token cookie.
        Requires valid authentication.
      operationId: logout
      responses:
        '204':
          description: Logout successful - no content returned
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
            Set-Cookie:
              description: Clears the refresh token cookie
              schema:
                type: string
                example: "refresh_token=; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=0"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/validate:
    get:
      tags:
        - Authentication
      summary: Validate authentication
      description: |
        Validates the current user's authentication status. 
        Useful for client applications to verify token validity without making other API calls.
      operationId: validateAuthentication
      responses:
        '200':
          description: Authentication is valid
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: string
              example:
                success: true
                message: "User authentication validated successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data: "User authentication validated successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: |
        Initiates a password reset process by sending a password reset link to the user's email address.
        
        **Security Note:** Returns success response regardless of whether the email exists in the system
        to prevent email enumeration attacks.
      operationId: forgotPassword
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset link sent (or would be sent if email exists)
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: "null"
              example:
                success: true
                message: "Password reset link has been sent successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data: null
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: |
        Resets a user's password using a valid password reset token received via email.
        The token must be valid and not expired (tokens expire after 60 minutes).
      operationId: resetPassword
      security: []  # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: "null"
              example:
                success: true
                message: "Password has been reset successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Image Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for API authentication
...
  headers:
    X-Rate-Limit-Remaining:
      description: Number of requests remaining in current rate limit window
      schema:
        type: integer
    X-Rate-Limit-Reset:
      description: Time when rate limit window resets (Unix timestamp)
      schema:
        type: integer
    X-Rate-Limit-Limit:
      description: Total number of requests allowed in rate limit window
      schema:
        type: integer
  schemas:
...
    ApiResponse:
      type: object
      discriminator:
        propertyName: success
        mapping:
          true: '#/components/schemas/SuccessResponse'
          false: '#/components/schemas/ErrorResponse'
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
        message:
          type: string
          description: Human-readable message describing the result
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the response was generated
      example:
        success: true
        message: "Operation completed successfully"
        timestamp: "2024-01-15T10:30:00Z"
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [true]
            data:
              description: The response payload containing the requested data
              oneOf:
                - type: object
                - type: array
                - type: string
                - type: number
                - type: boolean
                - type: "null"
      example:
        success: true
        message: "Data retrieved successfully"
        timestamp: "2024-01-15T10:30:00Z"
        data:
          id: 123
          name: "Example"
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - statusCode
            - statusText
            - code
          properties:
            success:
              type: boolean
              enum: [false]
            statusCode:
              type: integer
              description: HTTP status code
              minimum: 400
              maximum: 599
            statusText:
              type: string
              description: HTTP status text
            code:
              type: string
              description: Application-specific error code for programmatic handling
              enum:
                - NOT_FOUND
                - BAD_REQUEST
                - UNAUTHORIZED
                - ACCESS_DENIED
                - CONFLICT
                - FILE_STORAGE_ERROR
                - USER_NOT_FOUND
                - EMAIL_VERIFICATION_FAILED
                - PASSWORD_RESET_FAILED
                - MAIL_ERROR
                - INVALID_REQUEST
                - MISSING_PARAMETER
                - BUSINESS_ERROR
                - UNEXPECTED_ERROR
                - VALIDATION_ERROR
                - VALUE_ALREADY_EXISTS
                - RATE_LIMIT_EXCEEDED
                - IMAGE_LIMIT_EXCEEDED
                - EMAIL_SAME_AS_CURRENT
                - EMAIL_UPDATE_TOKEN_EXPIRED
                - EMAIL_UPDATE_TOKEN_INVALID
                - EMAIL_NO_LONGER_AVAILABLE
                - CURRENT_PASSWORD_INCORRECT
                - PASSWORD_SAME_AS_CURRENT
                - ROLE_UPDATE_SELF_DENIED
                - ROLE_UPDATE_ADMIN_DENIED
                - ADMIN_DISABLE_SELF_DENIED
                - ADMIN_CREATE_ADMIN_DENIED
                - INSUFFICIENT_TICKETS_AVAILABLE
      example:
        success: false
        message: "Resource not found"
        timestamp: "2024-01-15T10:30:00Z"
        statusCode: 404
        statusText: "Not Found"
        code: "NOT_FOUND"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required:
            - errors
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
            errors:
              type: object
              additionalProperties:
                type: string
              description: Field-specific validation errors mapped by field name
      example:
        success: false
        message: "Validation failed"
        timestamp: "2024-01-15T10:30:00Z"
        statusCode: 400
        statusText: "Bad Request"
        code: "VALIDATION_ERROR"
        errors:
          email: "INVALID_FORMAT"
          phoneNumber: "REQUIRED"

    # Pagination Schema
...
    PhoneNumberDTO:
      type: object
      required:
        - prefix
        - nationalNumber
      properties:
        prefix:
          type: string
          pattern: '^\\+\\d{1,3}'
          description: International phone number prefix (e.g., +1, +34, +44)
          example: "+1"
        nationalNumber:
          type: string
          pattern: '^\\d{1,14}$'
          description: National phone number without the international prefix
          example: "234567890"
      example:
        prefix: "+1"
        nationalNumber: "234567890"
    UserRegisterDTO:
      type: object
      required:
        - firstName
        - lastName
        - userName
        - email
        - phoneNumber
        - password
        - confirmPassword
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          description: User's first name
          example: "John"
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          description: User's last name
          example: "Doe"
        userName:
          type: string
          minLength: 2
          maxLength: 25
          description: Unique username for the user
          example: "johndoe"
        email:
          type: string
          format: email
          description: User's email address (used for login and notifications)
          example: "john.doe@animalrescue.org"
        phoneNumber:
          $ref: '#/components/schemas/PhoneNumberDTO'
        password:
          type: string
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%-^&_*(),.?":{}|<>]).{8,32}$'
          description: |
            Password must meet the following requirements:
            - 8-32 characters long
            - At least one lowercase letter
            - At least one uppercase letter  
            - At least one digit
            - At least one special character (!@#$%-^&_*(),.?":{}|<>)
          example: "SecurePass#123"
        confirmPassword:
          type: string
          description: Password confirmation (must match password)
          example: "SecurePass#123"
...
    RegisterAssociationData:
      type: object
      required:
        - associationName
        - addressData
      properties:
        associationName:
          type: string
          minLength: 2
          maxLength: 100
          description: Name of the animal rescue association
          example: "Happy Paws Animal Rescue"
        description:
          type: string
          maxLength: 500
          description: Brief description of the association's mission
          example: "Dedicated to rescuing and rehabilitating abandoned animals in our community"
        email:
          type: string
          format: email
          description: Association's contact email address
          example: "contact@happypaws.org"
        phoneNumber:
          $ref: '#/components/schemas/PhoneNumberDTO'
        addressData:
          $ref: '#/components/schemas/RegisterAddressData'
    RegisterRequest:
      type: object
      required:
        - userData
        - associationData
      properties:
        userData:
          $ref: '#/components/schemas/UserRegisterDTO'
        associationData:
          $ref: '#/components/schemas/RegisterAssociationData'
      example:
        userData:
          firstName: "John"
          lastName: "Doe"
          userName: "johndoe"
          email: "john.doe@animalrescue.org"
          phoneNumber:
            prefix: "+1"
            nationalNumber: "234567890"
          password: "SecurePass#123"
          confirmPassword: "SecurePass#123"
        associationData:
          associationName: "Happy Paws Animal Rescue"
          description: "Dedicated to rescuing and rehabilitating abandoned animals"
          email: "contact@happypaws.org"
          phoneNumber:
            prefix: "+1"
            nationalNumber: "555123456"
          addressData:
            placeId: "ChIJN1t_tDeuEmsRUsoyG83frY4"
            latitude: 40.7128
            longitude: -74.0060
            city: "New York"
            province: "NY"
            zipCode: "10001"
            formattedAddress: "123 Main St, New York, NY 10001, USA"
    RegisterResponse:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the created user
          example: 123
        email:
          type: string
          format: email
          description: The email address of the created user
          example: "john.doe@animalrescue.org"
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: User's email address or username
          example: "john.doe@animalrescue.org"
        password:
          type: string
          description: User's password
          example: "SecurePass#123"
    AuthResponse:
      type: object
      required:
        - accessToken
        - associationId
        - userId
      properties:
        accessToken:
          type: string
          description: JWT access token for API authentication
          example: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjMiLCJpYXQiOjE2MDk0NTkyMDB9.signature"
        associationId:
          type: integer
          format: int64
          description: The unique identifier of the user's association
          example: 456
        userId:
          type: integer
          format: int64
          description: The unique identifier of the authenticated user
          example: 123
    EmailVerificationRequest:
      type: object
      required:
        - verificationToken
      properties:
        verificationToken:
          type: string
          description: Email verification token received via email
          example: "550e8400-e29b-41d4-a716-446655440000"
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address of the account to reset password for
          example: "john.doe@animalrescue.org"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
        - confirmPassword
      properties:
        token:
          type: string
          description: Password reset token received via email
          example: "550e8400-e29b-41d4-a716-446655440000"
        password:
          type: string
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%-^&_*(),.?":{}|<>]).{8,32}$'
          description: |
            New password must meet the following requirements:
            - 8-32 characters long
            - At least one lowercase letter
            - At least one uppercase letter  
            - At least one digit
            - At least one special character (!@#$%-^&_*(),.?":{}|<>)
          example: "NewSecurePass#456"
        confirmPassword:
          type: string
          description: Password confirmation (must match password)
          example: "NewSecurePass#456"
...
  responses:
...
    BadRequest:
      description: Bad Request - The request was invalid or malformed
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid request parameters"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 400
            statusText: "Bad Request"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - Authentication required or token invalid
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 401
            statusText: "Unauthorized"
            code: "UNAUTHORIZED"
...
    Conflict:
      description: Conflict - Resource already exists or business rule violation
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource already exists"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 409
            statusText: "Conflict"
            code: "CONFLICT"
    ValidationError:
      description: Unprocessable Entity - Validation failed
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            message: "Validation failed"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 400
            statusText: "Bad Request"
            code: "VALIDATION_ERROR"
            errors:
              email: "INVALID_FORMAT"
              phoneNumber: "REQUIRED"
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Rate limit exceeded. Please try again later."
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 429
            statusText: "Too Many Requests"
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 500
            statusText: "Internal Server Error"
            code: "UNEXPECTED_ERROR" 